fix: hidden rainbow when nft avatar isn't exist (#5320)
(cherry picked from commit 2920d498d5b660d86967292872a5ea883882367f)
